requirements.txt
appdirs==1.4.3
Django==1.8.17
django-admin-bootstrapped==2.5.7
django-cors-headers==2.0.2
django-filter==1.0.1
django-location-field==2.0.2
django-money==0.10.2
django-subdomains==2.1.0
django-taggit==0.22.0
django-taggit-serializer==0.1.5
django-taggit-suggest==0.2
djangorestframework==3.6.2
geopy==1.11.0
Markdown==2.6.8
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2==2.7.1
py-moneyed==0.6.0
pyparsing==2.2.0
six==1.10.0
